Use trait to remove duplication
